+++ com.unity.cinemachine/Samples~/Shared Assets/Scripts/SamplesDynamicUI.cs
@@ -53,9 +53,9 @@ namespace Unity.Cinemachine.Samples
                 Debug.LogError("Cannot find TogglesAndButtons.  Is the source asset set in the UIDocument?");
                 return;
             }
-            
-            foreach (var item in Buttons)
+            for (int i = 0; i < Buttons.Count; ++i)
             {
+                var item = Buttons[i];
                 if (item.IsToggle.Enabled)
                 {
                     var toggle = new Toggle
@@ -87,8 +87,8 @@ namespace Unity.Cinemachine.Samples
 
         void OnDisable()
         {
-            foreach (var element in m_DynamicElements)
-                m_Root.Remove(element);
+            for (int i = 0; i < m_DynamicElements.Count; ++i)
+                m_Root.Remove(m_DynamicElements[i]);
             m_DynamicElements.Clear();
         }
     }