+++ com.unity.cinemachine/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added CinemachineCameraEvents and CinemachineBrainEvents behaviours for event processing.
 - Added BlendFinished and CameraDeactivated events.
 - Samples UI works with both built-in Input System.
+- New 2D platformer sample showing custom camera manager.
 
 ### Changed
 - Minimum Unity version is now 2022.2.15f1.